requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
easyocr==1.7.2
executing==2.2.0
filelock==3.17.0
fonttools==4.55.6
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langchain-core==0.3.33
langchain-openai==0.3.3
langsmith==0.3.4
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.1.1
openai==1.61.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.2
shapely==2.0.7
simplejwt==2.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.8.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
tracker==0.1.1
traitlets==5.14.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.85
ultralytics-thop==2.0.14
urllib3==2.3.0
virtualenv==20.31.2
wcwidth==0.2.13
whitenoise==6.9.0
zstandard==0.23.0
